# Implement groups accumulator for stddev and variance