… datafusion/sqllogictest/test_files/aggregate.slt