## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #12094.

## Rationale for this change

Hopefully improve performance of queries using stddev and variance aggregates.

I did not find any of the current benchmarks containing stddev and/or variance. So I modified clickbench query 31 and 32 to use STDDEV instead of AVG. And these were the results:

Before
```
Running benchmarks with the following options: RunOpt { query: None, common: CommonOpt { iterations: 5, partitions: None, batch_size: 8192, debug: false, string_view: false }, path: "/home/eejbyfeldt/dev/apache/datafusion/benchmarks/data/hits.parquet", queries_path: "/home/eejbyfeldt/dev/apache/datafusion/benchmarks/queries/clickbench/queries.sql", output_path: Some("/home/eejbyfeldt/dev/apache/datafusion/benchmarks/results/implement-groups-accumulator-for-stddev/clickbench_1.json") }
Q0: SELECT "SearchEngineID", "ClientIP", COUNT(*) AS c, SUM("IsRefresh"), STDDEV("ResolutionWidth") FROM hits WHERE "SearchPhrase" <> '' GROUP BY "SearchEngineID", "ClientIP" ORDER BY c DESC LIMIT 10;
Query 0 iteration 0 took 3945.8 ms and returned 10 rows
Query 0 iteration 1 took 3993.8 ms and returned 10 rows
Query 0 iteration 2 took 4016.9 ms and returned 10 rows
Query 0 iteration 3 took 3991.3 ms and returned 10 rows
Query 0 iteration 4 took 4024.2 ms and returned 10 rows
Q1: SELECT "WatchID", "ClientIP", COUNT(*) AS c, SUM("IsRefresh"), STDDEV("ResolutionWidth") FROM hits WHERE "SearchPhrase" <> '' GROUP BY "WatchID", "ClientIP" ORDER BY c DESC LIMIT 10;
Query 1 iteration 0 took 6551.8 ms and returned 10 rows
Query 1 iteration 1 took 6642.9 ms and returned 10 rows
Query 1 iteration 2 took 6618.6 ms and returned 10 rows
Query 1 iteration 3 took 6680.7 ms and returned 10 rows
Query 1 iteration 4 took 6568.3 ms and returned 10 rows
Done
```
After
```
Running benchmarks with the following options: RunOpt { query: None, common: CommonOpt { iterations: 5, partitions: None, batch_size: 8192, debug: false, string_view: false }, path: "/home/eejbyfeldt/dev/apache/datafusion/benchmarks/data/hits.parquet", queries_path: "/home/eejbyfeldt/dev/apache/datafusion/benchmarks/queries/clickbench/queries.sql", output_path: Some("/home/eejbyfeldt/dev/apache/datafusion/benchmarks/results/implement-groups-accumulator-for-stddev/clickbench_1.json") }
Q0: SELECT "SearchEngineID", "ClientIP", COUNT(*) AS c, SUM("IsRefresh"), STDDEV("ResolutionWidth") FROM hits WHERE "SearchPhrase" <> '' GROUP BY "SearchEngineID", "ClientIP" ORDER BY c DESC LIMIT 10;
Query 0 iteration 0 took 2144.0 ms and returned 10 rows
Query 0 iteration 1 took 2155.1 ms and returned 10 rows
Query 0 iteration 2 took 2095.3 ms and returned 10 rows
Query 0 iteration 3 took 2100.2 ms and returned 10 rows
Query 0 iteration 4 took 2238.7 ms and returned 10 rows
Q1: SELECT "WatchID", "ClientIP", COUNT(*) AS c, SUM("IsRefresh"), STDDEV("ResolutionWidth") FROM hits WHERE "SearchPhrase" <> '' GROUP BY "WatchID", "ClientIP" ORDER BY c DESC LIMIT 10;
Query 1 iteration 0 took 3075.6 ms and returned 10 rows
Query 1 iteration 1 took 3062.4 ms and returned 10 rows
Query 1 iteration 2 took 3081.9 ms and returned 10 rows
Query 1 iteration 3 took 3105.1 ms and returned 10 rows
Query 1 iteration 4 took 3101.6 ms and returned 10 rows
Done
```


<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

## What changes are included in this PR?

An implementation of GroupsAccumulator for the stddev and variance aggregates. 

It extracts the core logic from `NullState::acummulate` to a free function `accumulate` to make it reusable when implementing `GroupsAccumulators` that do not need separate null tracking as can be determine from some other value (in this case the count).

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

## Are these changes tested?
Added more test cases in aggregates.slt

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

## Are there any user-facing changes?
No.

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->