@andygrove 2024-08-22
Thanks @eejbyfeldt. Do you have any before/after benchmark results?