@eejbyfeldt 2024-08-23
> Thanks @eejbyfeldt. Do you have any before/after benchmark results?

Yes some benchmarks sounds like a great idea. I updated the PR description with something basic queries. 

FAs far as I could tell stddev/varaiance was not used in any current benchmarks, so I modified two clickbench queries to use stddev instead of avg. The queries were choosen based on them having big improvements when the GroupsAccumulator api was first implemented (according to https://www.influxdata.com/blog/aggregating-millions-groups-fast-apache-arrow-datafusion/) so they are probably showing best case improvement.